Phase 3.3: Extract RateLimiter class to rate_limiter.py
- Create doc_search/rate_limiter.py module with RateLimiter class
- RateLimiter provides thread-safe per-domain rate limiting:
  - Configurable default delay between requests
  - Custom per-domain delays
  - Backoff support for rate limiting responses

- Update crawler.py to import RateLimiter from rate_limiter
- Remove duplicate class definition from crawler.py
- Remove unused import (defaultdict no longer needed in crawler.py)

All 427 tests pass.
